GeoGals presentation for IAUS 377 (2023)
A presentation of results from the first and second GeoGals papers for a professional astronomy audience. Originally presented in Kuala Lumpur at the International Astronomy Union Symposium 377. The theme of this conference was "From the Milky Way to JWST - Connecting the near and the far". This was a lot of fun.